+++ ts/features/itwallet/discovery/components/ItwDiscoveryInfoComponent.tsx
@@ -38,11 +38,12 @@ import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender.ts";
 import { tosConfigSelector } from "../../../tos/store/selectors/index.ts";
 import { ITW_SCREENVIEW_EVENTS } from "../../analytics/enum.ts";
 import {
-  trackItWalletActivationStart,
-  trackItwDiscoveryPlus,
   trackItwIntroBack,
-  trackOpenItwTos
-} from "../../analytics/index.ts";
+  trackItWalletActivationStart,
+  trackItwDiscoveryPlus
+} from "../analytics";
+import { trackOpenItwTos } from "../../analytics";
+import { itwMixPanelCredentialDetailsSelector } from "../../analytics/store/selectors";
 import { useItwDismissalDialog } from "../../common/hooks/useItwDismissalDialog.tsx";
 import { itwIsActivationDisabledSelector } from "../../common/store/selectors/remoteConfig.ts";
 import { generateItwIOMarkdownRules } from "../../common/utils/markdown.tsx";
@@ -69,6 +70,9 @@ export const ItwDiscoveryInfoComponent = ({ credentialType }: Props) => {
   const itwActivationDisabled = useIOSelector(itwIsActivationDisabledSelector);
   const { tos_url } = useIOSelector(tosConfigSelector);
   const isWalletValid = useIOSelector(itwLifecycleIsValidSelector);
+  const mixPanelCredentialDetails = useIOSelector(
+    itwMixPanelCredentialDetailsSelector
+  );
   const toast = useIOToast();
 
   useOnFirstRender(
@@ -118,9 +122,9 @@ export const ItwDiscoveryInfoComponent = ({ credentialType }: Props) => {
   });
 
   const handleContinuePress = useCallback(() => {
-    trackItWalletActivationStart("L3");
+    trackItWalletActivationStart("L3", mixPanelCredentialDetails);
     machineRef.send({ type: "accept-tos" });
-  }, [machineRef]);
+  }, [machineRef, mixPanelCredentialDetails]);
 
   const [productHighlightsLayout, setProductHighlightsLayout] = useState({
     y: 0,