+++ ts/features/itwallet/issuance/screens/ItwIssuanceCredentialIntroductionScreen.tsx
@@ -8,10 +8,10 @@ import { useCallback, useMemo } from "react";
 import { useHeaderSecondLevel } from "../../../../hooks/useHeaderSecondLevel";
 import { useIOSelector } from "../../../../store/hooks";
 import {
-  getMixPanelCredential,
   trackItwCredentialIntro,
   trackItwCredentialStartIssuing
-} from "../../analytics";
+} from "../analytics";
+import { getMixPanelCredential } from "../../analytics/utils";
 import { itwLifecycleIsITWalletValidSelector } from "../../lifecycle/store/selectors";
 import { ItwCredentialIssuanceMachineContext } from "../../machine/credential/provider";
 import { getCredentialNameFromType } from "../../common/utils/itwCredentialUtils";