+++ ts/features/itwallet/machine/credential/actions.ts
@@ -7,14 +7,14 @@ import { checkCurrentSession } from "../../../authentication/common/store/action
 import { useIOStore } from "../../../../store/hooks";
 import { assert } from "../../../../utils/assert";
 import {
-  getMixPanelCredential,
-  trackAddCredentialProfileAndSuperProperties,
   trackSaveCredentialSuccess,
   trackStartAddNewCredential,
   trackStartCredentialUpgrade,
   trackWalletDataShare,
   trackWalletDataShareAccepted
 } from "../../analytics";
+import { getMixPanelCredential } from "../../analytics/utils";
+import { itwMixPanelCredentialDetailsSelector } from "../../analytics/store/selectors";
 import {
   itwCredentialsRemoveByType,
   itwCredentialsStore
@@ -156,10 +156,13 @@ export const createCredentialIssuanceActionsImplementation = (
     CredentialIssuanceEvents
   >) => {
     if (context.credentialType) {
-      const isItwL3 = itwLifecycleIsITWalletValidSelector(store.getState());
+      const state = store.getState();
+      const isItwL3 = itwLifecycleIsITWalletValidSelector(state);
       const credential = getMixPanelCredential(context.credentialType, isItwL3);
-      trackSaveCredentialSuccess(credential);
-      trackAddCredentialProfileAndSuperProperties(store.getState(), credential);
+      trackSaveCredentialSuccess({
+        credential,
+        credential_details: itwMixPanelCredentialDetailsSelector(state)
+      });
     }
   },