+++ ts/features/itwallet/machine/eid/actions.ts
@@ -12,9 +12,9 @@ import {
   trackItwDeactivated,
   trackItwIdAuthenticationCompleted,
   trackItwIdVerifiedDocument,
-  trackSaveCredentialSuccess,
-  updateITWStatusAndPIDProperties
+  trackSaveCredentialSuccess
 } from "../../analytics";
+import { itwMixPanelCredentialDetailsSelector } from "../../analytics/store/selectors";
 import {
   itwSetAuthLevel,
   itwFreezeSimplifiedActivationRequirements,
@@ -226,6 +226,12 @@ export const createEidIssuanceActionsImplementation = (
     });
   },
 
+  navigateToUpgradeCredentialsScreen: () => {
+    navigation.navigate(ITW_ROUTES.MAIN, {
+      screen: ITW_ROUTES.ISSUANCE.UPGRADE_CREDENTIALS
+    });
+  },
+
   closeIssuance: ({
     context
   }: ActionArgs<Context, EidIssuanceEvents, EidIssuanceEvents>) => {
@@ -320,15 +326,16 @@ export const createEidIssuanceActionsImplementation = (
   trackWalletInstanceCreation: ({
     context
   }: ActionArgs<Context, EidIssuanceEvents, EidIssuanceEvents>) => {
-    trackSaveCredentialSuccess(
-      context.level === "l3" ? "ITW_PID" : "ITW_ID_V2"
-    );
-    updateITWStatusAndPIDProperties(store.getState());
+    trackSaveCredentialSuccess({
+      credential: context.level === "l3" ? "ITW_PID" : "ITW_ID_V2",
+      ITW_ID_method: context.identification?.mode,
+      credential_details: itwMixPanelCredentialDetailsSelector(store.getState())
+    });
   },
 
   trackWalletInstanceRevocation: () => {
     const isItwL3 = itwLifecycleIsITWalletValidSelector(store.getState());
-    trackItwDeactivated(store.getState(), isItwL3 ? "ITW_PID" : "ITW_ID_V2");
+    trackItwDeactivated(isItwL3 ? "ITW_PID" : "ITW_ID_V2");
   },
 
   trackIdentificationMethodSelected: ({