+++ ts/features/itwallet/presentation/details/screens/ItwPresentationCredentialDetailScreen.tsx
@@ -18,13 +18,13 @@ import {
 import { useIODispatch, useIOSelector } from "../../../../../store/hooks.ts";
 import { usePreventScreenCapture } from "../../../../../utils/hooks/usePreventScreenCapture.ts";
 import { identificationRequest } from "../../../../identification/store/actions/index.ts";
+import { CREDENTIAL_STATUS_MAP } from "../../../analytics/utils/types.ts";
 import {
-  CREDENTIAL_STATUS_MAP,
-  getMixPanelCredential,
   trackCredentialDetail,
-  trackWalletCredentialShowFAC_SIMILE,
-  trackWalletCredentialShowTrustmark
-} from "../../../analytics";
+  trackWalletCredentialShowTrustmark,
+  trackWalletCredentialShowFAC_SIMILE
+} from "../analytics";
+import { getMixPanelCredential } from "../../../analytics/utils/index.ts";
 import ItwCredentialNotFound from "../../../common/components/ItwCredentialNotFound.tsx";
 import { PoweredByItWalletText } from "../../../common/components/PoweredByItWalletText.tsx";
 import { useItwFeaturesEnabled } from "../../../common/hooks/useItwFeaturesEnabled.ts";